fix: 이력서 조회에 대한 필터링 수정

+++ src/main/java/com/tave/tavewebsite/domain/resume/dto/response/ResumeEvaluateResDto.java
@@ -4,13 +4,13 @@ import org.springframework.data.domain.Page;
 
 public record ResumeEvaluateResDto(
         long totalRecruiter,
-        int notCompletedRecruiter,
-        int completedRecruiter,
+        long notCompletedRecruiter,
+        long completedRecruiter,
         Page<ResumeResDto> resumeResDtos
 ) {
     public static ResumeEvaluateResDto fromResume(long totalRecruiter,
-                                                  int notCompletedRecruiter,
-                                                  int completedRecruiter,
+                                                  long notCompletedRecruiter,
+                                                  long completedRecruiter,
                                                   Page<ResumeResDto> resumeResDtos) {
         return new ResumeEvaluateResDto(
                 totalRecruiter,
